Create swap.cpp
Here is the summary of the project:

Summary
In this lab, you complete a C++ program that swaps values stored in three int variables and determines maximum and minimum values. The C++ file provided for this lab contains the necessary variable declarations, as well as the input and output statements. You want to end up with the smallest value stored in the variable named first and the largest value stored in the variable named third. You need to write the statements that compare the values and swap them if appropriate. Comments included in the code tell you where to write your statements.

Instructions
Ensure the Swap.cpp file is open in your editor.
Write the statements that test the first two integers, and swap them if necessary.
Write the statements that test the second and third integer, and swap them if necessary.
Write the statements that test the first and second integers again, and swap them if necessary.
Execute the program by clicking the Run button at the bottom of the screen using the following sets of input values.
101    22    -23
630    1500  9
21    2    2